+++ Package.swift
@@ -10,9 +10,11 @@ let package = Package(
     products: [
         .library( name: "BatteryView", targets: ["BatteryView"])
     ],
-    dependencies: [],
+    dependencies: [
+        .package(url: "https://github.com/yonat/SweeterSwift", from: "1.0.4"),
+    ],
     targets: [
-        .target(name: "BatteryView", dependencies: [], path: "Sources")
+        .target(name: "BatteryView", dependencies: ["SweeterSwift"], path: "Sources")
     ],
     swiftLanguageVersions: [.v5]
 )